Update .lando.yml to alter how Robo is installed
Because of issues with how the manual fetch and install of Robo worked, replaced with a composer directive that installs Robo as a global package, and updated tooling to reflect executable path.

Ensures that latest 2.x composer is used and removes some deprecated xdebug configs.

Removes the global install of Drupal/Coder in favor of using project local.